backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.4.2
pydantic-settings>=2.0.3

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.5

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1

# Development
black>=23.10.1
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.1

# Monitoring and logging
structlog>=23.2.0
prometheus-client>=0.18.0

# Optional: Redis for caching
redis>=5.0.1 